requirements.txt
streamlit
pandas
openpyxl
pymupdf
streamlit-pdf-viewer
Pillow
google-generativeai==0.8.3
gspread
oauth2client

